watch: Use trimmed exe name rather than tectonic
